### 2018-08-31 #28 @labarba
👋  @wrightaprilm — Thank you for your submission to JOSE. I see that you describe this as an R package for visualizing phylogenetic trees, and that the package has been used in support of workshops and a university course.

As described, however, this submission appears to be unsuitable for publication on its own. I point you to the JOSE [Author Guidelines](http://jose.theoj.org/about#author_guidelines), which explains the type of submissions we consider.

> JOSE accepts two types of submissions: (1) computational learning modules, created as open educational resources (OER), and (2) open-source software, created as educational technology or infrastructure.

In the first type, what would fit is a written-out tutorial, or a learning module used within the graduate course you mention in the paper, which uses this package such that the interactive computational parts are essential to complete the learning objectives.

Note that the guidelines also say:

> Computational learning modules should be complete and immediately usable for self-learning or adoption by other instructors.

In the second type (software submissions), the JOSE scope covers educational technology or infrastructure software (e.g., auto graders, course-management solutions, etc.).

We have seen in the first few submissions some software that's meant to be used in the classroom, to support specific teaching contexts. We (the editors) have been discussing how to adapt to this demand, and are still working this out. But one thing we quickly agreed on is that this type of software can only be useful to other instructors (or self-learners) if it contains fully worked-out examples as part of the documentation.

Furthermore, the expectations for the JOSE paper are greater than shown in your submission. Note that a required section is the _Statement of Need_:

> A key component of the JOSE paper is a statement by the authors, explaining the contribution made by the submitted artifacts to computationally enabled teaching and learning, and describing how they might be used by others.

Further below, it says:

> The goal is that someone reading the JOSE paper has enough information to decide if they'd be interested in adoping the learnig module or software. Readers will want to know how the content/software has been used, and how they would adopt it. They may also want to be persuaded that the authors have put careful work on creating the material, and have experience teaching with it.

Please let me know what you want to do with the submission, given this feedback. For now, it's not ready to go into review.
