### 2018-08-31 #28 @labarba
I'm sorry I missed that issue conversation.

It looks like @moorepants was pointing you to make a submission that combines the software with the accompanying learning materials, together, as a learning module.

It's possible I'm missing something. I'm not an R person and I can't view the files directly on the repo, so I don't know if you have written-out learning materials there. If you do, perhaps the README and the paper need to explain how a reader would approach this.

Are you submitting learning materials, in addition to the package? If yes, how would I inspect the material?
