### 2018-08-31 #28 @moorepants
@labarba, @wrightaprilm submitted https://github.com/openjournals/jose/issues/20 for a pre-inquiry and I took it upon myself to respond.

I reviewed that inquiry and saw an educational software package, worksheets, and slides that looked like a reasonably complete set of teaching materials. Furthermore, the content was written as plain text sources. I told her it looked valid to submit based on these things. Her preliminary question was wrt to the fact that the submission was a mixture of both types of submissions we accept. I suggested submitting it as an educational software package and to treat the learning module materials as supplementary materials.

@wrightaprilm I apologize for leading you astray. I wrongly assumed that software packages designed to facilitate teaching and learning were valid. It seems that only things that are considered infrastructure are allowed. Maybe there are some things that can be added or improved to meet the submission criteria.
