### 2018-08-31 #28 @labarba
> I wrongly assumed that software packages designed to facilitate teaching and learning were valid. It seems that only things that are considered infrastructure are allowed. Maybe there are some things that can be added or improved to meet the submission criteria.

Please see above, where I stated: "We (the editors) have been discussing how to adapt to this demand, and are still working this out."

We are not ruling out these types of submissions. We were just a bit blindsided by them, as they weren't included originally.

In any case, full worked-out examples (a.k.a. tutorial) is essential.

I am not able to see slides or tutorial material on this submission. Perhaps we need breadcrumbs in the README. Also, the paper needs work, as I explained.
