### 2018-09-01 #28 @labarba
@wrightaprilm — Are you able to export your worksheets to a portable format, viewable on the repo? (like markdown). I don't have RStudio, so am not able to see this material. Also, the material is not described in the paper so I didn't get that it existed.
