### 2018-09-01 #28 @labarba
To clarify, the submission of _the worksheets_, together with the software, may be eligible under the "learning module" category.
