+++ vignettes/00-Instuctor-Guide.Rmd
@@ -1,0 +1,311 @@
+---
+title: "treesifter Instructor's Guide"
+output:
+  html_document: default
+  pdf_document: default
+bibliography: ../bibliography.bib
+---
+
+# Introduction
+
+treesiftr is an R package [@R] for visualizing the relationship between phylogenetic 
+trees and phylogenetic data. Phylogenetic trees are crucial to the study of 
+comparative biology, taxonomy, and evolution. However, understanding how to read 
+a phylogenetic tree, and how a phylogenetic tree relates to underlying 
+phylogenetic data, remains challenging.
+
+# How to Use the Instructor's Guide
+
+treesiftr has two modes. The first is a graphical user interface (GUI). Exercise 
+01-treesiftr is meant to be used with the graphical interface. The second is the RStudio interface. 02-treesiftr-advanced is meant to be used in RStudio, with 
+executeable code blocks.
+
+The first part of the instructor guide concerns the use of the GUI. The second part
+explains the RStudio interface, with worked examples.
+
+> *Tip*
+>
+> Linked text goes to the glossary. If you see a term you don't recognize, remember
+> you can refresh your memory at the bottom of this worksheet!
+
+
+
+## 01-treesiftr Instructor's Guide
+
+Navigate to the [treesiftr application](https://wrightaprilm.shinyapps.io/treesiftr_app/).
+
+treesiftr works by subsetting a phylogenetic matrix using the `start` 
+and `step` arguments. These are found on the left-hand side of the screen.
+The `start` argument controls where in the matrix you would like to begin visualizing characters. For example, a start value of 1 would indicate to begin visualizing characters from the first character in the matrix. The `step` value indicates how 
+many characters at once to visualize. A step value of three would indicate characters 
+should be viewed in threes. For example, if `start` = 1 and step = `3`, the visualization will show characters 1, 2, and 3. The start and end points of character visualizations are noted in the upper left-hand corner of the visualization.
+
+There are also two checkboxes. Leave these unchecked until otherwise noted.
+
+treesiftr works by subsetting a phylogenetic matrix by user input. Then, a [parsimony tree](#glossary) is constructed in Phangorn [@Schliep2011, Schliep2017] from the user-defined subset. The tree is [scored](#glossary) under both parsimony and Lewis' Mk model [@Lewis2001] for discrete character data. The data and tree are then visualized using ggtree [@ggtree], based upon the ggplot2 package [@ggplot2]. This application makes use of Shiny to provide a graphical interface, but there is a second included tutorial for more experienced users of the R statistical language.
+
+## Parsimony Questions
+
+1. Visualize characters 1 and 2. What is the parsimony score for this character set. Click "Do you want to print the [parsimony score?](#glossary)" in the interface to check your answer.
+
+> _Instructor's Note:_ The answer is two, as viewed in the below screenshot.
+
+![Question One Answer](img/Q1.png)
+
+2. Visualize characters 2 and 3. What [monophyletic](#glossary) group from tree 1 is no longer on this tree? Click below for the answer
+
+> _Instructor's Note:_ The answer is _Ursavus brevirhinus_, _Ballusia elmensis_,
+> and _Ursavus primaevus_. _Ursavus primaevus_ is no longer sister to _Ursavus     
+> brevirhinus_ and _Ballusia elmensis_ on the tree of character 2-3. Sample output 
+> follows.
+
+![Question One Answer](img/Q2.png)
+
+
+3. What is the parsimony score of the 31-34 character set? Click "Do you want to print the parsimony score?" in the infterface to check your answer.
+ 
+> _Instructor's Note:_ The answer is 7. The output you should see is below.
+![Question One Answer](img/Q3.png)
+
+4. Which character, 8, 9 or 10, represents a [reversal](#glossary)? 
+
[…]
