### 2018-09-02 54acdeb
various typos
