### 2018-09-02 c574a52
add slides as HTML
