### 2018-09-02 #28 @wrightaprilm
Thinking about viewing worked examples without RStudio: I made a [website](https://wrightaprilm.github.io/treesiftr/index.html) for the software + exercises, which renders both exercises under the "Articles" tab. There's an [instructor's guide](https://wrightaprilm.github.io/treesiftr/articles/00-Instuctor-Guide.html), with the output that students will see via the RStudio interface, as well as screenshots of the GUI exercise answers.

In re: the Paper. I was looking at the paper on my local machine, and I couldn't see the problem. Then I looked at the repository and realized the text had gotten lost in the rollback of a prior commit. Sorry about that. The text is [here](https://github.com/wrightaprilm/treesiftr/blob/master/paper.md). I'm not sure how changes come through the journal interface.

Supplemental slides - I saved these to [HTML](https://github.com/wrightaprilm/treesiftr/blob/master/inst/slides/TreeThinking.Rmd.html) and [PDF](https://github.com/wrightaprilm/treesiftr/blob/master/inst/slides/TreeThinking.pdf). I have to think a little more about how best to serve these outside of the RStudio application -  the PDFs look fine (though figure citations aren't appearing right), but the HTML slides look a little odd. I'm certainly willing to look more into this, if you think the module is OK to go out for review.

I'm just coming off maternity leave, and a bit slow on the uptake, so I need this next bit to summarize and keep me on track:

- [x] Slides viewable via Git repository.
- [x] Instructor's guide with worked examples that does not require RStudio to view.
- [x] Paper text actually present in paper (sorry about this again).
