+++ bibliography.bib
@@ -10330,5 +10330,28 @@ eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/evo.13117},
 abstract = {Evolutionary biologists have long predicted that evolutionary trade-offs among traits should constrain morphological divergence and species diversification. However, this prediction has yet to be tested in a broad evolutionary context in many diverse clades, including ants. Here, we reconstruct an expanded ant phylogeny representing 82\% of ant genera, compile a new family-wide trait database, and conduct various trait-based analyses to show that defensive traits in ants do exhibit an evolutionary trade-off. In particular, the use of a functional sting negatively correlates with a suite of other defensive traits including spines, large eye size, and large colony size. Furthermore, we find that several of the defensive traits that trade off with a sting are also positively correlated with each other and drive increased diversification, further suggesting that these traits form a defensive suite. Our results support the hypothesis that trade-offs in defensive traits significantly constrain trait evolution and influence species diversification in ants.}
 }
 
+
+@article{icytree,
+author = {Vaughan, T. G.},
+title = {IcyTree: Rapid browser-based visualization for phlogenetic trees and networks},
+journal = {Bioinformatics},
+year    = {2017},
+pages = {315-328},
+doi = {10.1093/bioinformatics/btx155},
+}
+
+@article{phylogenyio,
+ title = {Interactive web-based visualization of phylogenetic trees using Phylogeny.IO},
+ author = {Jovanovic, Nikola and Mikheyev, Alexander S},
+ year = {2016},
+ volume = {4},
+ pages = {e2579v1},
+ journal = {PeerJ Preprints},
+ issn = {2167-9843},
+ url = {https://doi.org/10.7287/peerj.preprints.2579v1},
+ doi = {10.7287/peerj.preprints.2579v1}
+}
+
+
 @Comment{jabref-meta: databaseType:bibtex;}
 
