### 2018-10-08 #28 @labarba
:wave @juanklopper — I've assigned this submission to you as handling editor. Since this will be your first editing assignment, let me know if you need to bounce ideas about selecting reviewers. Check out the [Editor's Guide](https://github.com/labarba/jose/blob/master/EditorsGuide.md) I put together: it describes the editor workflow, including needed `@whedon` commands.

@wrightaprilm — My apologies for the delay in handling this submission! The beginning of semester hit me like a hurricane.
