### 2018-10-12 #28 @wrightaprilm
@labarba  I'm in grant writing jail, and hadn't noticed the passage of time. Thanks for your help handling this!

Let me know what you need from me and when you need it. I'm happy to provide some suggested reviewers who are working at the intersection of evolution and education.
