### 2018-10-16 #28 @juanklopper
@wrightaprilm I was originally assigned as editor for your submission.  Unfortunately, it has taken me this long to be in a position to assist you in taking this proposal forward.  Thanks to our editor-in-chief (EiC), @labarba for helping out.  Not to mention @moorepants .  Thanks everyone.

I note that you are in grant writing jail!  It is a special place isn't it!? :-)  I wish you all the luck.

Now that I have had some time to look at your work, I have some suggestions and questions.  Let's start with the latter.  Your last post above mentioned that you wanted to suggest a couple of reviewers.  Have you progressed with this?

@labarba has intimated to the fact that we are still negotiating the perils of inclusion criteria for the journal, and quite rightly so.  At this time, though, I want to support your submission.  Creating learning resources that others can use is something that is close to my heart.  I want to suggest that you have a look at one of the papers in our journal:  Pynamical: Model and visualize discrete nonlinear dynamical systems, chaos, and fractals.  It is available [here](https://github.com/gboeing/pynamical) .  Much like your project, it is a novel software package (in python), combined with a well developed resource infrastructure that makes it ideal for someone else to pick up and use as an educational resource.

To this end, I love the website that you added and can only suggest that you expand on it a little bit.  The same goes for the great Shiny app.  From viewing your work submitted for this proposal, I sense that your are an expert in your field.  Something, I suppose, not all of your colleagues are.  I would really recommend then that you expand on both the website and Shiny app.  Just to make it a _pick-up-and-use_ for the _less fortunate_.

I also want to suggest a quick rewrite of you actual paper.  I come from a world where we have very structured papers.  As our EiC once mentioned, the papers for JOSE are just the entry-portal to your work (for our readers).  I wouldn't mind, though, if you would consider the following structure for the rewrite:

**Summary**

A summary of the R package (one paragraph).

A summary of the R Shiny app (one paragraph).

A summary of the accompanying resources detailing the use (how to) of the package / app / accompanying resources (one paragraph).

Introduction summary on phylogenetics (one to two paragraphs).

**Statement of need**

Explain the need for this R package / Shiny app / accompanying educational resources within the context of its use (one paragraph).

Explain a use-case of the software / learning resources with inclusion of screenshots.

Let me know what you think.  And now, I'll go read up on the phylogenetics of trees, because you made me interested in it.
