### 2018-10-17 #28 @wrightaprilm
Hi Juan-

Happy to do a restructure; I'll probably get to it Friday afternoon. I send off one of two grants Friday morning, which means I get to do something fun:)

I'm happy to suggest reviewers. I have two other primarily undergraduate institution faculty in mind (one an R expert, one not) - I'll send along names and emails when I commit the restructure.
