+++ paper.md
@@ -17,36 +17,17 @@ bibliography: bibliography.bib
 
 ## Summary
 
-Phylogenetic trees represent the evolutionary relationships between a set of taxa. Estimating 
-phylogenetic trees is crucial in many areas of biology. However, visualizing the 
-relationship between a phylogenetic tree and the data used to generate it is not always 
-intuitive to novice learners. To assist with understanding this relationship, I have created treesiftr, an R 
-package [@R] that creates subsets of data from a phylogenetic matrix, estimates a 
-tree for each subset under the maximum parsimony optimality criterion, and scores that 
-tree under both the likelihood and parsimony criteria. The output of the package is a 
-visualization or set of visualizations of a tree and the characters used to estimate it. 
-Also included is a (Shiny application](https://wrightaprilm.shinyapps.io/treesiftr_app/) [@shiny] 
-that can be run locally or used via the web without installing any R packages or 
-having knowledge of R.
+treesiftr is an R package [@R], and associated Shiny [@shiny] [application]((https://wrightaprilm.shinyapps.io/treesiftr_app/) for visualizing the relationship between phylogenetic trees and the underlying data used to estimate them. treesiftr works by subsetting a phylogenetic matrix according to user-provided input about
+which characters to visualize. A maximum parsimony tree is then estimated from each dataset subset. Maximum parsimony was chosen for speed and analytical simplicity. Under the parsimony optimality criterion, the preferred tree is the one that suggests the fewest evolutionary steps, or character changes over evolutionary history. The tree is scored under both parsimony and  Lewis' Mk model [@Lewis2001], a maximum likelihood model for estimating phylogeny from 
+discrete character data. The data and tree are then visualized using ggtree [@ggtree], based upon the ggplot2 [@ggplot2] package. Expected outputs are the same whether the learner is interacting via the GUI or the RStudio interface; however, the RStudio interface does have additional options not available in the GUI. 
 
-In the `vignettes` directory, there are two worksheets for student use. One worksheet
-(01-treesiftr) is intended for use with the Shiny server graphical user interface (GUI). The other,
-(02-treesiftr-advanced) is intended for use with R and RStudio. There is also an
-instructor's guide with screenshots of the expected GUI output, and worked code snippets and outputs
-for the advanced exercise. These exercises can also be viewed as a [rendered website](https://wrightaprilm.github.io/treesiftr/),
-so that no installations or downloads are required at all for the GUI-based exercises.
+The included Shiny [application]((https://wrightaprilm.shinyapps.io/treesiftr_app/) renders the visual output of the subsetting and estimation process, and can be used to provide further input to the treesiftr functions, such as decorating the trees with scores under different optimality criteria. It can be run locally or used via the web without installing any R packages or having knowledge of R.
 
-treesiftr works by subsetting a phylogenetic matrix according to user-provided input about
-which characters to visualize. A maximum parsimony tree is then estimated from each dataset subset. 
-Maximum parsimony was chosen for speed and analytical simplicity. Under the parsimony
-optimality criterion, the preferred tree is the one that suggests the fewest evolutionary steps,
-or character changes over evolutionary history. The tree is scored under both parsimony and 
-Lewis' Mk model [@Lewis2001], a maximum likelihood model for estimating phylogeny from 
-discrete character data. The data and tree are then visualized using ggtree [@ggtree],
-based upon the ggplot2 [@ggplot2] package. Expected outputs are the same whether the learner is interacting via the GUI or the 
-RStudio interface; however, the RStudio interface does have additional options not available in the GUI.
-The worksheet for the RStudio interface also emphasizes general R skills, such as subsetting data, 
-and specific R skills, such as interacting with phylogenetic data.
[…]
