+++ bibliography.bib
@@ -2661,6 +2661,30 @@ Coalescence Theory, Population Size, Divergence Times},
   Timestamp                = {2009.02.23}
 }
 
+@article{feng,
+  title={Progressive sequence alignment as a prerequisitetto correct phylogenetic trees},
+  author={Feng, Da-Fei and Doolittle, Russell F},
+  journal={Journal of molecular evolution},
+  volume={25},
+  number={4},
+  pages={351--360},
+  year={1987},
+  publisher={Springer}
+}
+
+@article{FarrisKluge,
+author = {Farris, James S. and Kluge, Arnold G. and Eckardt, Michael J.},
+title = {A Numerical Approach to Phylogenetic Systematics},
+journal = {Systematic Biology},
+volume = {19},
+number = {2},
+pages = {172-189},
+year = {1970},
+doi = {10.2307/2412452},
+URL = {http://dx.doi.org/10.2307/2412452},
+eprint = {/oup/backfile/content_public/journal/sysbio/19/2/10.2307/2412452/2/19-2-172.pdf}
+}
+
 @Article{Fitch1971,
   Title                    = {Toward Defining the Course of Evolution: Minimum Change for a Specific Tree Topology},
   Author                   = {Fitch, Walter M.},
@@ -3412,6 +3436,16 @@ Uses also the diagnosis of the golden runs.},
   Publisher                = {SMBE},
   Timestamp                = {2010.10.26}
 }
+
+@article{gower1969,
+  title={Minimum spanning trees and single linkage cluster analysis},
+  author={Gower, John C and Ross, Gavin JS},
+  journal={Applied statistics},
+  pages={54--64},
+  year={1969},
+  publisher={JSTOR}
+}
+
 
 @PhdThesis{Gowri-Shankar2006,
   Title                    = {RNA PHYLOGENETIC INFERENCE WITH HETEROGENEOUS SUBSTITUTION MODELS},
