### 2018-10-21 ec07434
more updats to paper
