### 2018-10-24 766424a
adding edits to paper
