### 2018-10-24 #28 @wrightaprilm
Alright, I committed changes to the paper that are more in line with what you've suggested. Let me know what you think.
