### 2018-10-29 #28 @juanklopper
@wrightaprilm I have read your rewrite.  Thanks a lot.  Just to keep you informed, we have now moved on with the process and have approached some potential reviewers (I do note that you might have some suggestions and we'll keep this in mind).  Hopefully, we can update you shortly.
