### 2018-10-29 #28 @wrightaprilm
I'd like to suggest [Cathy Newman](https://www.catherinenewman.org/contact), who is another primarily undergrad institution faculty member serving similar educational audiences. [Rowan Lockwood](https://www.wm.edu/as/geology/people/faculty/lockwood_r.php) is another one who would be good.
