### 2018-11-04 #35 @whedon
[REVIEW]: treesiftr: An R package and server for viewing phylogenetic trees and data
**Submitting author:** @wrightaprilm (<a href="http://orcid.org/0000-0003-4692-3225">April Wright</a>)
**Repository:** <a href="https://github.com/wrightaprilm/treesiftr" target ="_blank">https://github.com/wrightaprilm/treesiftr</a>
**Version:** v1.0.0

- [x] **Version:** Does the release version given match the repository release (v1.0.0)?

- [x] **Version:** Does the release version given match the repository release (v1.0.0)?
