### 2018-11-04 #35 @labarba
@ethanwhite, @rachelss — Thank you for agreeing to review this submission for JOSE! We are happy to welcome you to our adventure in scholarly publishing of open-source educational software and learning resources.

JOSE has published seven papers so far 😃 — have a look: http://jose.theoj.org
Check our [Reviewer Guidelines](http://jose.theoj.org/about#reviewer_guidelines) and let us know if you have any questions.

@juanklopper will be your handling editor: you're in good hands.
