### 2018-11-05 #35 @labarba
Thanks for reporting this, @ethanwhite. I agree that this doesn't present conflict.
