### 2018-11-05 #35 @ethanwhite
A couple of questions/comments for @wrightaprilm from the "General checks" section:

> License: Does the repository contain a plain-text LICENSE file with the contents of a standard license? (OSI-approved for code, Creative Commons for content)

Following the R packaging standard the `LICENSE` file does not include the actual license, only the year and copyright holder. The license is given in the `DESCRIPTION` file. One common approach to dealing with the conflicts between R's packaging standard and the more general open source standard of having the full license in the `LICENSE` file is to have the full license in `LICENSE.md` (which is included in `.Rbuildignore`) and keep `LICENSE` in the standard R format. This is rOpenSci's general strategy. I'm not sure what JOSS or JOSE's general approach is to this issue.

>  Version: Does the release version given match the repository release (v1.0.0)?

The repository is lacking a current release. If you're waiting to tag `1.0.0` until other comments have been satisfied that makes sense and I'll wait until it looks like things are wrapping up and then ping this issue again about tagging a release.
