+++ CONDUCT.md
@@ -1,0 +1,25 @@
+As contributors and maintainers of this project, we pledge to respect all people who contribute through reporting issues, 
+posting feature requests, updating documentation, submitting pull requests or patches, and other activities.
+
+treesiftr is dedicated to providing a harassment-free community experience for all participants 
+regardless of gender, gender identity and expression, sexual orientation, disability, 
+physical appearance, body size, race, age or religion, familial status, or socioeconomic status. 
+The developer, contributors, and users do not tolerate harassment in any form in communications
+related to this project, whether those communications are public (comments on issues or pull requests,
+discussions at conferences), or private (e-mails or direct messages)
+
+Examples of unacceptable behavior by participants include the use of sexual language or 
+imagery, derogatory comments or personal attacks, trolling, public or private harassment, 
+insults, or other unprofessional conduct. 
+
+Project maintainers have the right and responsibility to remove, edit, or reject comments, 
+commits, code, wiki edits, issues, and other contributions that are not aligned to this 
+Code of Conduct. Project maintainers also reserve the right to block users who refuse to
+follow the policies of this document. Project maintainers who do not follow the Code of 
+Conduct may be removed from the project team.
+
+Instances of abusive, harassing, or otherwise unacceptable behavior may be reported by 
+opening an issue or contacting maintainer [April Wright](wright.aprilm@gmail.com) directly.
+
+This Code of Conduct is adapted from the Contributor Covenant, version 1.0.0, 
+available at https://www.contributor-covenant.org/version/1/0/0/code-of-conduct.html
