### 2018-11-08 8a17c3d
update .travis
