+++ DESCRIPTION
@@ -8,7 +8,7 @@ Description: R package for building parsimony trees from subsets of characters i
 	and visualizing the characters alongside the representations of the tree.
 License: MIT + file LICENSE
 Depends:
-	shiny
+    shiny
 Imports:
     phangorn,
     ggtree,
@@ -21,4 +21,4 @@ Suggests:
     covr
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 6.0.1
+RoxygenNote: 6.1.0
