+++ .travis.yml
@@ -6,8 +6,5 @@ cache: packages
 warnings_are_errors: FALSE
 bioc_required: yes
 use_bioc: true
-r_packages:
-- covr
-- testthat
 after_success:
     - Rscript -e 'covr::codecov()'
