+++ R/data.R
@@ -2,7 +2,15 @@
 #'
 #' A dataset containing 18 taxa and 62 characters for bears.
 #'
-#' @format A fasta dataset frame with 53940 rows and 10 variables
-#' @format A phylo object of a phylogeny
+#' @docType data
+#' @keywords datasets
+#' @aliases phy_mat tree
+#' @usage data(bears)
+#' @format A collection of objects representing function inputs:
+#' \describe{
+#'     \item{phy_mat}{A phylogenetic matrix with 62 characters and 18 taxa}
+#'     \item{tree}{A phylogenetic tree to which tips will be added}
+#' }
 #' @source \url{https://github.com/revbayes/revbayes_tutorial/tree/master/RB_TotalEvidenceDating_FBD_Tutorial/data}
+
 "bears"
