+++ vignettes/00-Instuctor-Guide.Rmd
@@ -1,8 +1,11 @@
 ---
 title: "treesifter Instructor's Guide"
-output:
-  html_document: default
-  pdf_document: default
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteEngine{knitr::knitr}
+  %\VignetteIndexEntry{Instructor's Guide}
+  %\usepackage[UTF-8]{inputenc}
 bibliography: ../bibliography.bib
 ---
 
