+++ NAMESPACE
@@ -4,3 +4,14 @@ export(check_subs)
 export(generate_sliding)
 export(generate_tree_vec)
 export(generate_tree_vis)
+export(tree_dat)
+importFrom(ape,as.phylo)
+importFrom(ape,multi2di)
+importFrom(ggplot2,ggtitle)
+importFrom(ggtree,ggtree)
+importFrom(ggtree,msaplot)
+importFrom(phangorn,acctran)
+importFrom(phangorn,phyDat)
+importFrom(phangorn,pml)
+importFrom(utils,head)
+importFrom(utils,tail)
