### 2018-11-15 #35 @labarba
@rachelss — Our dear bot `whedon` generates the PDF for us right here in this issue. Scroll up and you will see:
