### 2018-12-12 #35 @juanklopper
@wrightaprilm I'm back from all my duties as external examiner and wanted to know how you are getting along.  Big thanks to @rachelss  Excellent insight and suggestions.
