### 2018-12-12 #35 @wrightaprilm
I think we're good, thanks! Just waiting on @ethanwhite,  in case there are conflicts between the two reviews.

Which is fine, I think we've all had some end-of-semester stuff going on that would make revising hard ;)
