### 2018-12-12 #35 @ethanwhite
Here's my review. Thanks for your patience and understanding @wrightaprilm. The end of the semester is indeed an... interesting time.

Overall, great work and really useful. I think this will be a great addition to a number of workshops and classrooms. For every checkbox that I haven't checked I've provided a description of what I think can be improved to make this ready to go.

## General Checks

* License: The license file follows R's guidelines, but as a result the repository does not include an actual license.
* Version: The versions do not currently match because v1.0.0 does not currently exists. The author plans to release this version once the rest of the review is complete.

## Documentation

* Usage: There are examples of how to use the module and questions that can be asked of students along with the answers. These are both very helpful. I think it would be useful in both the README and the Instructions guide to include some documentation designed to explain specifically how someone would adopt the module in a teaching setting. This is covered a bit in the JOSE paper and you get the idea that the learning material is lab exercise focused from the vignettes, but I'd suggest also putting some of those ideas front and center in the README and Instructors Guide so that folks more immediately understand the goals of the package and learing materials.
* I'd also suggest adding a pdf of the Instructors Guide and linking to it directly from the README so that it is easy to find and read for someone first encountering the package online.
* Community guildelines: There is a code of conduct (which is awesome!) but I didn't find information on how to contribute or seek support.

## Pedagogy / Instructional design

* Learning Objectives: are a little difficult to determine from the current documentation. I understand that they are about helping students understand tree construction and phylogeny more broadly, but it would be helpful to lay this out a little more explicitly. I think a second short paragraph in the Introduction of the Instructors guide should be sufficient.

## JOSE paper

* Use in classroom or other settings: I think just a couple of additional sentences describing how to adopt this material would make a big differences for helping make clear to readers exactly how they could use the material.
* Could someone else teach with this module: I think the module needs just a little more documentation. Once the changes described above are made I think this criteria will be satisfied.
* References: Lewis 2001 is missing the DOI: 10.1080/106351501753462876
