### 2019-01-02 cd04619
adding documentation fixes in response to reviews
