+++ inst/shiny-examples/app.R
@@ -1,3 +1,0 @@
-library(shiny)
-
-shinyApp(ui = ui, server = server)
