+++ bibliography.bib
@@ -800,6 +800,22 @@ journal = {Proceedings of the National Academy of Sciences}
   Timestamp                = {2008.09.17}
 }
 
+@article {Baum979,
+	author = {Baum, David A. and Smith, Stacey DeWitt and Donovan, Samuel S. S.},
+	title = {The Tree-Thinking Challenge},
+	volume = {310},
+	number = {5750},
+	pages = {979--980},
+	year = {2005},
+	doi = {10.1126/science.1117727},
+	publisher = {American Association for the Advancement of Science},
+	issn = {0036-8075},
+	URL = {http://science.sciencemag.org/content/310/5750/979},
+	eprint = {http://science.sciencemag.org/content/310/5750/979.full.pdf},
+	journal = {Science}
+}
+
+
 @Article{Baum1992,
   Title                    = {Combining Trees as a Way of Combining Data Sets for Phylogenetic Inference, and the Desirability of Combining Gene Trees},
   Author                   = {Baum, Bernard R.},
@@ -5542,7 +5558,7 @@ Their algorithms are available in BAMBE (Bayesian Analysis in Molecular Biology 
   Number                   = {6},
   Pages                    = {913-925},
   Volume                   = {50},
-
+  DOI					   = {10.1080/106351501753462876},
   Owner                    = {pett},
   Publisher                = {Oxford University Press},
   Timestamp                = {2016.12.23}
