+++ README.md
@@ -1,0 +1,194 @@
+![logo](inst/images/treesiftR.png) [![Travis build
+status](https://travis-ci.org/wrightaprilm/treesiftr.svg?branch=master)](https://travis-ci.org/wrightaprilm/treesiftr)
+[![Coverage
+status](https://codecov.io/gh/wrightaprilm/treesiftr/branch/master/graph/badge.svg)](https://codecov.io/github/wrightaprilm/treesiftr?branch=master)
+
+Introduction
+------------
+
+Estimating phylogenetic trees is crucial in many areas of evolutionary
+biology. However, visualizing the relationship between data and trees is
+not intutive. To assist with visualizing this relationship, I have
+created [treesiftR](https://wrightaprilm.github.io/treesiftr/), an R
+package (R Core Team 2018) that takes subsets of data from a
+phylogenetic matrix, generates a tree under parsimony, and scores that
+tree under both the likelihood and parsimony criteria. The output of the
+package is a visualization or set of visualizations of a tree and
+characters. Also included is a [Shiny
+application](https://wrightaprilm.shinyapps.io/treesiftr_app/) (Chang et
+al. 2018) that can be run locally or used via the web without installing
+any R packages or having knowledge of R.
+
+If you are interested in treesiftr, take a look at my [instructor’s
+guide](https://wrightaprilm.github.io/treesiftr/articles/00-Instructor-Guide.html)
+for more information (PDF also available
+[here](https://github.com/wrightaprilm/treesiftr/blob/master/vignettes/00-Instuctor-Guide.pdf))
+about adopting this module.
+
+### Target Audience
+
+treesiftr has been used in the [Analytical Paleobiology
+Workshop](http://www.analytical.palaeobiology.de/), in which the
+audience was graduate students and postdocs, many of whom had no prior
+knowledge of phylogeny. It is also used in the Genetics course at
+Southeastern Louisiana University, where the audience is undergraduates
+who have no prior knowledge of phylogeny. It is meant to be accompanied
+by lecture material on phylogenetics. A glossary is provided with each
+worksheet, and a sample slide deck is included in the `inst/slides`
+directory.
+
+Installation
+------------
+
+Currently, treesiftr can be installed via the devtools `install_github`
+function (Wickham, Hester, and Chang 2018).
+
+``` r
+devtools::install_github("wrightaprilm/treesiftr")
+```
+
+Required Packages
+-----------------
+
+``` r
+knitr::opts_chunk$set(echo = TRUE)
+library(ape)
+library(treesiftr)
+library(phangorn)
+library(alignfigR)
+```
+
+    ## Welcome to alignfigR!
+
+``` r
+library(ggtree)
+```
+
+    ## ggtree v1.13.2  For help: https://guangchuangyu.github.io/software/ggtree
+    ## 
+    ## If you use ggtree in published research, please cite:
+    ## Guangchuang Yu, David Smith, Huachen Zhu, Yi Guan, Tommy Tsan-Yuk Lam. ggtree: an R package for visualization and annotation of phylogenetic trees with their covariates and other associated data. Methods in Ecology and Evolution 2017, 8(1):28-36, doi:10.1111/2041-210X.12628
+
+    ## 
+    ## Attaching package: 'ggtree'
+
+    ## The following object is masked from 'package:ape':
+    ## 
+    ##     rotate
+
+``` r
+library(ggplot2)
+```
+
+Operation
+---------
+
+The first step to making a treesiftr visualization is to select the
[…]
