### 2019-01-02 e1a6650
remove function in progress
