+++ NAMESPACE
@@ -1,7 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
 export(check_subs)
-export(generate_anc_states)
 export(generate_sliding)
 export(generate_tree_vec)
 export(generate_tree_vis)
