### 2019-01-04 eebcb25
adding random addition tree infrastructure
