+++ README.Rmd
@@ -1,6 +1,8 @@
 ---
 title: "README.Rmd"
-output: html_document
+output:
+  md_document:
+    variant: markdown_github
 bibliography: bibliography.bib
 ---
   ![logo](inst/images/treesiftR.png)
@@ -79,6 +81,18 @@ output_vector
 ```
 
 `Phangorn` [@Schliep2011, Schliep2017] requires a starting tree to estimate a parsimony tree. We specify the tree we read in earlier for this purpose. The trees, which were generated with `ggtree` [@ggtree], a ggplot2 [@ggplot2] library for phylogenies, have been saved to a vector, which can be displayed in its entirety, or subsetted to look at specific trees.
+
+## Random Trees
+
+You can also map the characters to a random tree, and count the parsimony steps 
+or score the characters under likelihood on this tree.
+
+
+```{r message=FALSE,  warning=FALSE}
+output_vector <- generate_tree_vis(sample_df = sample_df, alignment =                                                     aln_path, tree = tree, phy_mat = bears,
+                             random_tree = TRUE)
+output_vector
+```
 
 
 ## References
