+++ NAMESPACE
@@ -1,12 +1,14 @@
 # Generated by roxygen2: do not edit by hand
 
 export(check_subs)
+export(generate_anc_states)
 export(generate_sliding)
 export(generate_tree_vec)
 export(generate_tree_vis)
 export(tree_dat)
 importFrom(ape,as.phylo)
 importFrom(ape,multi2di)
+importFrom(ape,rtree)
 importFrom(ggplot2,ggtitle)
 importFrom(ggtree,ggtree)
 importFrom(ggtree,msaplot)
