### 2019-01-04 ed334ad
refactor for appropriate return
