### 2019-01-04 c53b093
fix test in reflection of new function
