+++ tests/testthat/test_vis_vec.R
@@ -9,6 +9,4 @@ test_that("Test generate_tree_vis(data_frame, alignment, tree, pscore, lscore)",
   vec <- generate_tree_vis(sample_df = sample_df, alignment = aln_path,
                            tree = tree, phy_mat = bears)
     testthat::expect_equal(length(vec), 5)
-    testthat::expect_equal(length(vec[[1]]$layers), 5)
-    testthat::expect_is(vec[[1]]$labels$title, "character")
 })
