### 2019-01-05 #35 @ethanwhite
LGTM! Pending editorial input on the license (my personal take on the best compromise solution is above in the discussion related to adding a `LICENSE.md` file) and the version being updated to `1.0.0` I think this is ready to go. Nice work @wrightaprilm!
