### 2019-01-05 #35 @labarba
👋 @rachelss — It sounds like the author has addressed a bunch of review comments. Can you go over things and let us know here if your comments have been addressed and you are ready to recommend publication? Thanks!
