### 2019-01-05 #35 @labarba
@wrightaprilm You have an extra paren that breaks the link in the beginning of the second paragraph.
