### 2019-01-05 #35 @rachelss
I looked. I'm sorry, but I do have more comments, just not right this second. I was trying to figure out exactly how to explain the challenges I would still have in teaching from the given instructions and material.
