### 2019-01-07 #35 @rachelss
"This course is a 30-day paleobiological data" - there's a typo here

shinyapps.io limits you to 25 hours / month - if you have more than 50 undergrads using this app for class I forsee a problem that instructors should be aware of before attempting to use the app to teach

Remind users how to install ggtree in the readme (it's a bioconductor package not cran)

need to fix link to instructor's guide in readme

typo in Parsimony Questions (01-treesiftr.pdf) q3

Reorder questions to group LOs

Need to tell students to uncheck "show parsimony score" after q1

q2 is difficult and confusing - I was looking for a monophyletic clade of a genus, which isn't the case and leaves me asking questions about what the data are telling me. I also have a hard time switching back and forth between the two trees and remember what was where.

Now you can add a question about parsimony scores on random trees and students can see reversals / repeat substitutions for a suboptimal tree

Are q1 and q3 essentially the same question? Does this serve a purpose? If the goal is to see a sub that happens in two different parts of the tree ask students about this.

q4 - prompt students to view the tree estimated from 8-10 first

q6 - I'm not sure what to do here. Do you need to prompt students to look at several different datasets that have the same parsimony score? Do you mean characters or alignments?

q7 - Since you don't have branch lengths on the tree I'm not sure how students are supposed to know that changes are occurring on longer or shorter branches.

q9 - Technically the answer to this q is 0 so you might want to clarify that you're only using PI sites.

q10 - point students to a specific set of characters

I expect most users will want to teach with the app so maybe point toward that first and tell people not to worry about the code.

Hide the Rmd output for the libraries in the readme.

In the readme you assign two different things to sample_df

Maybe just show generating one tree in the readme - you jump straight to instructions for generating a batch of trees from a batch of datasets.

To get at @ethanwhite request for more clarity on the use of the module I suggest looking at the literature on tree-thinking, particularly concept inventories. This work would provide motivation for teachers to use a tool to address specific thought processes that students find challenging. You need to point out to potential users that students really struggle to understand certain concepts and this app provides a hands-on way for them to work on these concepts.

Also, do you picture this as a lab / recitation? homework? small class? large class? one question at a time during a lecture? I'm picturing this as a lab/recitation with a TA circulating to provide help, but curious what you're thinking.

Sorry this is a lot for a second round - the first time I was struggling with the interface so didn't look at the questions. I'm ok with the Shiny app now there is more guidance on what I'm looking at.
