### 2019-01-08 #35 @wrightaprilm
No need to apologize - these are all great notes. I'll try to get them turned around tomorrow!
