### 2019-01-10 c287bc7
editing readme and vignettes in response to reviews
