### 2019-01-10 9798bc9
update website and knit vignettes
