+++ vignettes/00-Instuctor-Guide.Rmd
@@ -47,10 +47,9 @@ explains the RStudio interface, with worked examples.
 
 - Be able to count parsimony changes for the displayed characters on 
 the estimated tree.
-  - Questions: 1, 3
-- Apply phylogenetic terminology, such as reversal, autapomorphy, and monophyletic,
-to discuss trait and lineage evolution.
-  - Questions: 2, 4, 5
+  - Questions: 1, 2
+- Read a phylogenetic tree. Using the information on the tree, apply phylogenetic terminology such as reversal, autapomorphy, and monophyletic to discuss trait and lineage evolution.
+  - Questions: 3, 4, 5
 - Explain why two characters with the same parsimony score may not have the same 
 likelihood score, and how the number of characters changes the likelihood and 
 parsimony scores.
@@ -79,22 +78,21 @@ treesiftr works by subsetting a phylogenetic matrix by user input. Then, a [pars
 
 ![Question One Answer](img/Q1.png)
 
-2. Visualize characters 2 and 3. What [monophyletic](#glossary) group from tree 1 is no longer on this tree? 
+2. Visualize characters 1 and 2. This time, click the "View random set of trees rather than an estimated tree" button in the interface. Are the parsimony scores larger or smaller than those of the character on an estimated tree? Does this make sense to you? Why? Remember to uncheck
+this box before continuing. 
+ 
+> _Instructor's Note:_ They should be larger. Parsimony estimation will favor the shortest tree
+while no such criteria is applied to a random tree. Sample output is below, though the 
+exact numbers and distribution will vary.
+![Question Two Answer](img/Q2.png)
+
+3. Visualize characters 2 and 3. What [monophyletic](#glossary) group from tree 1 is no longer on this tree? 
 
 > _Instructor's Note:_ The answer is _Ursavus brevirhinus_, _Ballusia elmensis_,
 > and _Ursavus primaevus_. _Ursavus primaevus_ is no longer sister to _Ursavus     
 > brevirhinus_ and _Ballusia elmensis_ on the tree of character 2-3. Sample output 
 > follows.
 
-![Question Two Answer](img/Q2.png)
-
-
-3. Visualize characters 1 and 2. This time, click the "View random set of trees rather than an estimated tree" button in the interface. Are the parsimony scores larger or smaller than those of the character on an estimated tree? Does this make sense to you? Why? Remember to uncheck
-this box before continuing. 
- 
-> _Instructor's Note:_ They should be larger. Parsimony estimation will favor the shortest tree
-while no such criteria is applied to a random tree. Sample output is below, though the 
-exact numbers and distribution will vary.
 ![Question Three Answer](img/Q3.png)
 
 4. Which character, 8, 9 or 10, represents a [reversal](#glossary)? 
@@ -244,14 +242,7 @@ output_vector
 > _Instructor's Note:_ The answer is two, as viewed in the first visualization in
 the vector.
 
-2. Visualize character set 2-3 and 3-4. What [monophyletic](#glossary) group from the 
-tree for characters 2-3 is no longer on the tree for characters 3-4? 
-
-> _Instructor's Note:_ The answer is _Ursavus brevirhinus_ and _Ballusia elmensis_.
-This clade is broken up on the tree for characters 3-4, in which  _Ballusia elmensis_
-is sister to _Zaragocyon_daamsi_.
-
-3. Visualize characters 1 and 2. This time, click the "View random set of trees rather than an estimated tree" button in the interface. Are the parsimony scores larger or smaller than those of the character on an estimated tree? Does this make sense to you? Why?
[…]
