+++ bibliography.bib
@@ -27,7 +27,7 @@
 
 @Article{abella12,
   Title                    = {\textit{{K}retzoiarctos} gen. nov., the Oldest Member of the Giant Panda Clade},
-  Author                   = {Juan Abella and David M. Alba and Josep M. Robles and Alberto Valenciano and Cheyenn Rotgers and Ra\{ul Carmona and Plinio Montoya and Jorge Morales},
+  Author                   = {Juan Abella and David M. Alba and Josep M. Robles and Alberto Valenciano and Cheyenn Rotgers and Ra\"ul Carmona and Plinio Montoya and Jorge Morales},
   Journal                  = {PLoS One},
   Year                     = {2012},
   Pages                    = {e48985},
@@ -44,22 +44,8 @@
   Volume                   = {67}
 }
 
-@article {Alvarez-Carretero441105,
-	author = {Alvarez-Carretero, Sandra and Goswami, Anjali and Yang, Ziheng and dos Reis, Mario},
-	title = {Bayesian Estimation of Species Divergence Times Using Correlated Quantitative Characters},
-	elocation-id = {441105},
-	year = {2018},
-	doi = {10.1101/441105},
-	publisher = {Cold Spring Harbor Laboratory},
-	abstract = {Discrete morphological data have been widely used to study species evolution, but the use of quantitative (or continuous) morphological characters is less common. Here, we implement a Bayesian method to estimate species divergence times using quantitative characters. Quantitative character evolution is modelled using Brownian diffusion with character correlation and accounting for character variation within populations. Through simulations, we demonstrate that ignoring the population variation (or population {noise{) and the correlation among characters leads to biased estimates of divergence times and rate, especially if the correlation and population noise are strong. We apply our new method to the analysis of quantitative characters (cranium landmarks) and molecular data from carnivoran mammals. Our results show that time estimates are affected by whether the correlations and population noise are accounted for or ignored in the analysis. The estimates are also affected by the type of data analysed, with analyses of morphological characters only, molecular data only, or a combination of both; showing noticeable differences among the time estimates. Rate variation of morphological characters among the carnivoran species appears to be very high, with Bayesian model selection indicating that the independent-rates model fits the morphological data better than the autocorrelated-rates model. We suggest that using morphological continuous characters, together with molecular data, can bring a new perspective to the study of species evolution. Our new model is implemented in the MCMCtree computer program for Bayesian inference of divergence times.},
-	URL = {https://www.biorxiv.org/content/early/2018/10/11/441105},
-	eprint = {https://www.biorxiv.org/content/early/2018/10/11/441105.full.pdf},
-	journal = {bioRxiv}
-}
-
-
 @article{beast2,
-    author = {Bouckaert, Remco and Heled, Joseph and K{\{u}hnert, Denise and Vaughan, Tim and Wu, Chieh-Hsi and Xie, Dong and Suchard, Marc A. and Rambaut, Andrew and Drummond, Alexei J.},
+    author = {Bouckaert, Remco and Heled, Joseph and K{\"u}hnert, Denise and Vaughan, Tim and Wu, Chieh-Hsi and Xie, Dong and Suchard, Marc A. and Rambaut, Andrew and Drummond, Alexei J.},
     journal = {PLOS Computational Biology},
     publisher = {Public Library of Science},
     title = {BEAST 2: A Software Platform for Bayesian Evolutionary Analysis},
@@ -227,7 +213,7 @@ url = {https://besjournals.onlinelibrary.wiley.com/doi/abs/10.1111/j.2041-210X.2
 }  
 @Article{abella12,
[…]
