### 2019-01-10 #35 @rachelss
Thank you for your patience with my extensive review. I think it's good to go. I hope to get a chance to use the Shiny app with undergrads - I think it's a really useful tool and having a set of questions ready to go is very helpful. Well done!

fyi - Please rebuild the Instructor's Guide pdf.

also fyi - I'm picturing this done with learnr to combine questions and app with instant feedback, but that's for the future
