### 2019-01-12 #35 @labarba
@rachelss — You have some unchecked items in the review checklist. Please check these off, when ready, and give me your recommendation to accept and publish, at that point.

@ethanwhite — You're missing the license checkmark. As for the version, after @wrightaprilm ups the version on the submission repository, for archival, we can use `@whedon set vx.x.x as version` here, to match.
