### 2019-01-13 #35 @ethanwhite
@labarba - My license checkmark is missing due to what I think is the still unresolved issue related to the fact that CRAN's approach to licensing does not include having the license text itself in the repository. I've suggested a solution to this and @wrightaprilm was requested editorial guidance. If JOSE is OK with not including the actual license text anywhere in the repository then I'm happy to check the box.
