### 2019-01-13 #35 @labarba
Thank you, @ethanwhite.

@wrightaprilm : Ethan said:

> One common approach to dealing with the conflicts between R's packaging standard and the more general open source standard of having the full license in the `LICENSE` file is to have the full license in `LICENSE.md` (which is included in `.Rbuildignore`) and keep `LICENSE` in the standard R format.

Could you do that?
