### 2019-01-14 #35 @rachelss
I checked off everything but license and version. Once those are resolved I think this should be published.
