### 2019-01-15 608311f
addressing last comments
