### 2019-01-15 de7eaf3
upping version number and fixing paper typo
