### 2019-01-15 #35 @labarba
@ethanwhite, @rachelss — The author has addressed remaining items. Could you check off the list and confirm your recommendation to publish?

@juanklopper — When both reviewers recommend publication, next step will be for @wrightaprilm to create an archive in Zenodo and report the DOI here. You can then use `@whedon set <doi> as archive`, and add the "accepted" label. I'll then publish!
