### 2019-01-16 #35 @labarba
@wrightaprilm — Please now make an archive on Zenodo, and report the DOI here. Make sure to edit the metadata as needed to match the paper (title and authors), as Zenodo just grabs that info from the repo.
