### 2019-01-16 #35 @wrightaprilm
Thanks so much for all your help, everyone! I've never made a Zenodo before - is [this](https://zenodo.org/record/2541824) right? I had to hand-edit some things, as a user who made a PR showed up as an author.
