### 2019-01-16 2980a3d
update bib file
