+++ bibliography.bib
@@ -10402,6 +10402,107 @@ doi = {10.1093/bioinformatics/btx155},
  doi = {10.7287/peerj.preprints.2579v1}
 }
 
+@Article{Lartillot2004,
+  author  = {N. Lartillot and H. Philippe},
+  title   = {A {B}ayesian mixture model for across-site heterogeneities in the amino-acid replacement process},
+  journal = {Molecular Biology and Evolution},
+  year    = {2004},
+  volume  = {21},
+  pages   = {1095--1109}
+}
+
+@article{blanchard,
+author = {Blanchard, Benjamin D. and Moreau, Corrie S},
+title = {Defensive traits exhibit an evolutionary trade-off and drive diversification in ants},
+journal = {Evolution},
+volume = {71},
+number = {2},
+year    = {2017},
+pages = {315-328},
+keywords = {Ancestral state reconstruction, defense, evolutionary trade-off, Formicidae, trait-based diversification},
+doi = {10.1111/evo.13117},
+url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/evo.13117},
+eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/evo.13117}
+}
+
+@article{icytree,
+author = {Vaughan, T. G.},
+title = {IcyTree: Rapid browser-based visualization for phlogenetic trees and networks},
+journal = {Bioinformatics},
+year    = {2017},
+pages = {315-328},
+doi = {10.1093/bioinformatics/btx155}
+}
+
+@article{phylogenyio,
+ title = {Interactive web-based visualization of phylogenetic trees using Phylogeny.IO},
+ author = {Jovanovic, Nikola and Mikheyev, Alexander S},
+ year = {2016},
+ volume = {4},
+ pages = {e2579v1},
+ journal = {PeerJ Preprints},
+ issn = {2167-9843},
+ url = {https://doi.org/10.7287/peerj.preprints.2579v1},
+ doi = {10.7287/peerj.preprints.2579v1}
+}
+
+@article{rudolph1998,
+  title={Evolution and the nature of science: On the historical discord and its implications for education},
+  author={Rudolph, John L and Stewart, Jim},
+  journal={Journal of Research in Science Teaching: The Official Journal of the National Association for Research in Science Teaching},
+  volume={35},
+  number={10},
+  pages={1069--1089},
+  year={1998},
+  publisher={Wiley Online Library}
+}
+
+@article{sandvik2008,
+  title={Tree thinking cannot taken for granted: challenges for teaching phylogenetics},
+  author={Sandvik, Hanno},
+  journal={Theory in Biosciences},
+  volume={127},
+  number={1},
+  pages={45--51},
+  year={2008},
+  publisher={Springer}
+}
+
+@article{ohara1997,
+  title={Population thinking and tree thinking in systematics},
+  author={O'HARA, ROBERT J},
+  journal={Zoologica scripta},
+  volume={26},
+  number={4},
+  pages={323--329},
+  year={1997},
+  publisher={Wiley Online Library}
+}
+
+@Article{Meisel2010,
+author={Meisel, Richard P.},
+title={Teaching Tree-Thinking to Undergraduate Biology Students},
+journal={Evolution: Education and Outreach},
+year={2010},
+month={Dec},
+day={01},
+volume={3},
+number={4},
+pages={621--628},
+issn={1936-6434},
+doi={10.1007/s12052-010-0254-9},
+url={https://doi.org/10.1007/s12052-010-0254-9}
+}
+
+@article{meir2007,
+author = { Eli  Meir,Judy  Perry,Jon C.  Herron,Joel  Kingsolver},
[…]
