### 2019-01-16 #35 @wrightaprilm
Oops, forgot to push my updated .bib file. And because all my references in the vignettes and website build locally before pushing, this is the first step where that would be caught. I've pushed it now. Sorry all!
