+++ paper.md
@@ -33,7 +33,7 @@ Once homology is assigned, a phylogeny is estimated from the data. Several optim
 
 ## Statement of Need
 
-Understanding phylogenetic trees is challenging for students [@Baum979, @Meisel2010]. Some of this challenge comes from inherently misunderstanding evolution in pre-Darwinian terms, such as a Platonic 'great chain of being' or as being ladder-like [@rudolph1998, @sandvik]. This is often inherently coupled to the idea of humans being the pinnacle of evolution [@ohara1997]. treesfitr, therefore, by default uses a non-human dataset to decouple tree-thinking from preconceived notions of organismal hierarchy. Students also have difficulty simply reading phylogenetic trees - often reading along the tips of a tree, as opposed to looking at the internal nodes [@Baum979, @Meisel2010]. To counter this misconception, treesiftr continuously estimates new phylogenies, moving both tips and nodes. This causes students to have to continually re-evaluate the tree and the information on it [@meir2007].
+Understanding phylogenetic trees is challenging for students [@Baum979, @Meisel2010]. Some of this challenge comes from inherently misunderstanding evolution in pre-Darwinian terms, such as a Platonic 'great chain of being' or as being ladder-like [@rudolph1998, @sandvik2008]. This is often inherently coupled to the idea of humans being the pinnacle of evolution [@ohara1997]. treesfitr, therefore, by default uses a non-human dataset to decouple tree-thinking from preconceived notions of organismal hierarchy. Students also have difficulty simply reading phylogenetic trees - often reading along the tips of a tree, as opposed to looking at the internal nodes [@Baum979, @Meisel2010]. To counter this misconception, treesiftr continuously estimates new phylogenies, moving both tips and nodes. This causes students to have to continually re-evaluate the tree and the information on it [@meir2007].
 
 There are many  phylogenetic tree viewers on the market such as [FigTree](http://tree.bio.ed.ac.uk/software/figtree/),
 IcyTree [@icytree], and Phylogeny.IO [@phylogenyio]. But visualizing the data that underlie a particular tree is still largely accomplished via the Mesquite software [@Maddison2008]. To use Mesquite, one must perform local installs. The software must also be interacted with via a GUI, without the opportunity for learners to practice programmatic skills while learning about phylogenetic trees and data. I wrote this package to allow students to learn about phylogenetic trees and data without performing local installs, and to give the option for students to practice programmatic skills while doing so. 
