### 2019-01-16 #35 @labarba
Hmm ... I think we ran into a funky bug. @arfon — What could be going on? `@whedon accept` fails, but `@whedon generate pdf` does compile the paper.
