+++ bibliography.bib
@@ -1,30 +1,4 @@
 % Encoding: windows-1252
-
-@String { AN        = {The American Naturalist} }
-@String { AREES     = {Annual Review of Ecology, Evolution, and Systematics} }
-@String { ARG       = {Annual Review of Genetics} }
-@String { BI        = {Bioinformatics} }
-@String { BMCBIO    = {BMC Biology} }
-@String { BMCEVOBIO = {BMC Evolutionary Biology} }
-@String { EL        = {Ecology Letters} }
-@String { EVOLUTION = {Evolution} }
-@String { GENETICS  = {Genetics} }
-@String { JASA      = {Journal of the American Statistical Association} }
-@String { JME       = {Journal of Molecular Evolution} }
-@String { JTB       = {Journal of Theoretical Biology} }
-@String { MBE       = {Molecular Biology and Evolution} }
-@String { MEE       = {Methods in Ecology and Evolution} }
-@String { MPE       = {Molecular Phylogenetics and Evolution} }
-@String { PBIO      = {PLoS Biology} }
-@String { PCB       = {PLoS Computational Biology} }
-@String { PGENETICS = {PLoS Genetics} }
-@String { PNAS      = {Proceedings of the National Academy of Sciences} }
-@String { PONE      = {PLoS One} }
-@String { ProcB     = {Proceedings of the Royal Society B: Biological Sciences} }
-@String { SYSBIO    = {Systematic Biology} }
-@String { TEE       = {Trends in Ecology \& Evolution} }
-@String { TPB       = {Theoretical Population Biology} }
-
 @Article{abella12,
   Title                    = {\textit{{K}retzoiarctos} gen. nov., the Oldest Member of the Giant Panda Clade},
   Author                   = {Juan Abella and David M. Alba and Josep M. Robles and Alberto Valenciano and Cheyenn Rotgers and Ra\"ul Carmona and Plinio Montoya and Jorge Morales},
@@ -149,33 +123,6 @@ url = {https://besjournals.onlinelibrary.wiley.com/doi/abs/10.1111/j.2041-210X.2
   Timestamp                = {2009.02.23}
 }
 
-% Encoding: windows-1252
-
-
-@String { AN        = {The American Naturalist} }
-@String { AREES     = {Annual Review of Ecology, Evolution, and Systematics} }
-@String { ARG       = {Annual Review of Genetics} }
-@String { BI        = {Bioinformatics} }
-@String { BMCBIO    = {BMC Biology} }
-@String { BMCEVOBIO = {BMC Evolutionary Biology} }
-@String { EL        = {Ecology Letters} }
-@String { EVOLUTION = {Evolution} }
-@String { GENETICS  = {Genetics} }
-@String { JASA      = {Journal of the American Statistical Association} }
-@String { JME       = {Journal of Molecular Evolution} }
-@String { JTB       = {Journal of Theoretical Biology} }
-@String { MBE       = {Molecular Biology and Evolution} }
-@String { MEE       = {Methods in Ecology and Evolution} }
-@String { MPE       = {Molecular Phylogenetics and Evolution} }
-@String { PBIO      = {PLoS Biology} }
-@String { PCB       = {PLoS Computational Biology} }
-@String { PGENETICS = {PLoS Genetics} }
-@String { PNAS      = {Proceedings of the National Academy of Sciences} }
-@String { PONE      = {PLoS One} }
-@String { ProcB     = {Proceedings of the Royal Society B: Biological Sciences} }
-@String { SYSBIO    = {Systematic Biology} }
-@String { TEE       = {Trends in Ecology \& Evolution} }
-@String { TPB       = {Theoretical Population Biology} }
 @Article{ape,
[…]
