### 2019-01-16 #35 @wrightaprilm
I think this is actually my issue. I use this bibtex on multiple multiple-author projects, and we include some string expansions for journals we cite from a bunch. I removed the translation block.
