### 2019-01-16 999292f
replace string expansion block
