### 2019-01-16 #35 @wrightaprilm
That was surprising! Not even a little bit what I expected to occur.

I added back in the string expansion block. There were two of them. It's compiling on local via R's Knitr machinery.
