### 2019-01-16 #35 @labarba
It compiles that way ... I notice that two references bleed out of the right margin (long DOIs). Can you think of any way to fix that? @arfon, we may still need your help here, sorry!
