### 2019-01-16 #35 @wrightaprilm
It's such an odd format for a DOI. I also see a render issue in the references on a hyphenated first name. I'll wait to push my fix until I hear back on this issue.
