### 2019-01-16 #35 @arfon
> It compiles that way ... I notice that two references bleed out of the right margin (long DOIs). Can you think of any way to fix that? @arfon, we may still need your help here, sorry!

Lemme see what I can do. I think https://github.com/openjournals/whedon/pull/39 might fix this.
