### 2019-01-16 #35 @arfon
Unfortunately this is a non-trivial fix. I can fix one of the DOI strings but the longer one is going to require some work. @labarba - it's up to you how we proceed from here.
