### 2019-01-16 #35 @labarba
@wrightaprilm Another option is for you to use the initial only for the second author and also abbreviate the journal title, in the hopes the DOI moves left enough.

Abbreviation from: http://images.webofknowledge.com/images/help/WOS/A_abrvjt.html
AMERICAN BIOLOGY TEACHER
	AM BIOL TEACH
