### 2019-01-16 #35 @wrightaprilm
Let me play with it a bit tomorrow. If I clone [Whedon](https://github.com/openjournals/whedon), I can compile within template on my local, correct? That way I can goof around without bothering everyone?

Thanks for all your help - what a weird little hiccup!
