### 2019-01-16 wrightaprilm/treesiftr#5 @labarba
Missing "and" between authors
I think the missing "and" between authors is causing an incomplete author list in the reference.

https://github.com/wrightaprilm/treesiftr/blob/999292f2720414f5abc27d88540c1c5dcac45c54/bibliography.bib#L10471
