### 2019-01-16 #35 @labarba
Please note this issue I just opened: https://github.com/wrightaprilm/treesiftr/issues/5
