### 2019-01-16 #35 @labarba
@wrightaprilm — You can use the command `@whedon generate pdf` here, yourself, to see how your changes affect the compiled paper.
