### 2019-01-16 411375f
fix missing and, and hyphenated first name issue
