+++ bibliography.bib
@@ -2583,7 +2583,7 @@ Coalescence Theory, Population Size, Divergence Times},
 
 @Article{Felsenstein1978,
   Title                    = {The Number of Evolutionary Trees},
-  Author                   = {Felsenstein, Joseph},
+  Author                   = {Felsenstein, J.},
   Journal                  = {Systematic Zoology},
   Year                     = {1978},
   Number                   = {1},
@@ -2652,7 +2652,7 @@ Coalescence Theory, Population Size, Divergence Times},
 
 @article{feng,
   title={Progressive sequence alignment as a prerequisitetto correct phylogenetic trees},
-  author={Feng, Da-Fei and Doolittle, Russell F},
+  author={Feng, D. and Doolittle, R. F.},
   journal={Journal of molecular evolution},
   volume={25},
   number={4},
@@ -5928,7 +5928,7 @@ url="https://doi.org/10.1007/s00285-009-0260-0"
 @Misc{Maddison2008,
   Title                    = {Mesquite: a modular system for evolutionary analysis. Version 2.5},
 
-  Author                   = {Maddison, W. P. and D.R. Maddison. .},
+  Author                   = {Maddison, W. P. and Maddison, D. R.},
   HowPublished             = {http://mesquiteproject.org},
   Year                     = {2008},
 
@@ -10468,7 +10468,7 @@ url={https://doi.org/10.1007/s12052-010-0254-9}
 }
 
 @article{meir2007,
-author = { Eli  Meir,Judy  Perry,Jon C.  Herron,Joel  Kingsolver},
+author = {Meir, E. Perry, J., Herron, J. C., and Kingsolver, J.},
 title = {College Students' Misconceptions About Evolutionary Trees},
 journal = {The American Biology Teacher},
 volume = {69},
