### 2019-01-17 #35 @arfon
> @arfon — What if we publish it, wait for the fix, then fix the PDF and update the Crossref deposit? The downside is remembering to do this — I find it ugly to have a published paper with a broken layout.

Yes, I think we should do this.
