+++ bibliography.bib
@@ -10473,7 +10473,7 @@ title = {College Students' Misconceptions About Evolutionary Trees},
 journal = {The American Biology Teacher},
 volume = {69},
 year = {2007},
-doi = {10.1662/0002-7685(2007)69[71:CSMAET]2.0.CO;2}
+doi = {https://doi.org/10.1662/0002-7685(2007)69[71:csmaet]2.0.co;2}
 }
 
 
