### 2019-01-17 7dd0624
remove duplicate citation
