+++ bibliography.bib
@@ -4737,15 +4737,6 @@ Gives also a introduction what else can be used and what is good and bad in ML, 
   Timestamp                = {2010.10.20}
 }
 
-@article{Jovanovic2016,
-  title={Interactive web-based visualization of phylogenetic trees using Phylogeny. IO},
-  author={Jovanovic, Nikola and Mikheyev, Alexander S},
-  year=2016,
-  journal={PeerJ Preprints},
-  volume={4},
-  pages={e2579v1},
-  publisher={PeerJ Inc. San Francisco, USA}
-}
 
 @Article{Jow2002,
   Title                    = {{Bayesian Phylogenetics Using an RNA Substitution Model Applied to Early Mammalian Evolution}},
@@ -10371,8 +10362,7 @@ doi = {10.1093/bioinformatics/btx155},
  pages = {e2579v1},
  journal = {PeerJ Preprints},
  issn = {2167-9843},
- url = {https://doi.org/10.7287/peerj.preprints.2579v1},
- doi = {10.7287/peerj.preprints.2579v1}
+ doi = {https://doi.org/10.7287/peerj.preprints.2579v1},
 }
 
 @Article{Lartillot2004,
@@ -10405,17 +10395,6 @@ journal = {Bioinformatics},
 year    = {2017},
 pages = {315-328},
 doi = {10.1093/bioinformatics/btx155}
-}
-
-@article{phylogenyio,
- title = {Interactive web-based visualization of phylogenetic trees using Phylogeny.IO},
- author = {Jovanovic, Nikola and Mikheyev, Alexander S},
- year = {2016},
- volume = {4},
- pages = {e2579v1},
- journal = {PeerJ Preprints},
- issn = {2167-9843},
- doi = {https://doi.org/10.7287/peerj.preprints.2579v1}
 }
 
 @article{rudolph1998,
