+++ bibliography.bib
@@ -10362,7 +10362,7 @@ doi = {10.1093/bioinformatics/btx155},
  pages = {e2579v1},
  journal = {PeerJ Preprints},
  issn = {2167-9843},
- doi = {https://doi.org/10.7287/peerj.preprints.2579v1},
+ doi = {10.7287/peerj.preprints.2579v1},
 }
 
 @Article{Lartillot2004,
