### 2019-01-17 #35 @wrightaprilm
All right, I zapped one of them. There's still a tiny bit of template breaking on the phylogeny.io citation, but I'm happy enough with it. I'd be fine to go ahead, then fix it later.
