### 2019-01-17 #35 @labarba
Feng & Doolittle (1987) has a DOI (not listed): https://doi.org/10.1007/BF02603120
(also, please capitalize the journal name)

Felsenstein (1973) does,  too:
https://doi.org/10.1093/sysbio/22.3.240

Felsenstein (1978) is
https://doi.org/10.2307/2412810

Gower & Ross (1969) is http://doi.org/10.2307/2346439

... please go down the reference list and add any more missing DOIs.
