+++ bibliography.bib
@@ -2587,6 +2587,7 @@ Coalescence Theory, Population Size, Divergence Times},
   Journal                  = {Systematic Zoology},
   Year                     = {1978},
   Number                   = {1},
+  DOI                      = {https://doi.org/10.2307/2412810},
   Pages                    = {27--33},
   Volume                   = {27},
 
@@ -2608,7 +2609,7 @@ Coalescence Theory, Population Size, Divergence Times},
   Number                   = {3},
   Pages                    = {240--249},
   Volume                   = {22},
-
+  DOI                      = {https://doi.org/10.1093/sysbio/22.3.240},
   Owner                    = {shhn001},
   Publisher                = {JSTOR},
   Timestamp                = {2008.04.02}
@@ -2651,12 +2652,13 @@ Coalescence Theory, Population Size, Divergence Times},
 }
 
 @article{feng,
-  title={Progressive sequence alignment as a prerequisitetto correct phylogenetic trees},
+  title={Progressive sequence alignment as a prerequisite to correct phylogenetic trees},
   author={Feng, D. and Doolittle, R. F.},
-  journal={Journal of molecular evolution},
+  journal={Journal of Molecular Evolution},
   volume={25},
   number={4},
   pages={351--360},
+  doi = {https://doi.org/10.1007/BF02603120},
   year={1987},
   publisher={Springer}
 }
@@ -3432,6 +3434,8 @@ Uses also the diagnosis of the golden runs.},
   journal={Applied statistics},
   pages={54--64},
   year={1969},
+  Doi                      = {10.2307/2346439},
+
   publisher={JSTOR}
 }
 
@@ -4406,7 +4410,7 @@ Gives also a introduction what else can be used and what is good and bad in ML, 
   Number                   = {8},
   Pages                    = {754--755},
   Volume                   = {17},
-
+  DOI                      = {https://doi.org/10.1093/bioinformatics/17.8.754},
   Owner                    = {shhn001},
   Publisher                = {Oxford Univ Press},
   Timestamp                = {2008.04.02}
@@ -10405,6 +10409,7 @@ doi = {10.1093/bioinformatics/btx155}
   number={10},
   pages={1069--1089},
   year={1998},
+  DOI={https://doi.org/10.1002/(SICI)1098-2736(199812)35:10<1069::AID-TEA2>3.0.CO;2-A},
   publisher={Wiley Online Library}
 }
 
@@ -10416,6 +10421,7 @@ doi = {10.1093/bioinformatics/btx155}
   number={1},
   pages={45--51},
   year={2008},
+  DOI={10.1007/s12064-008-0022-3},
   publisher={Springer}
 }
 
@@ -10427,6 +10433,7 @@ doi = {10.1093/bioinformatics/btx155}
   number={4},
   pages={323--329},
   year={1997},
+  doi={https://doi.org/10.1111/j.1463-6409.1997.tb00422.x},
   publisher={Wiley Online Library}
 }
 
