+++ bibliography.bib
@@ -3431,7 +3431,7 @@ Uses also the diagnosis of the golden runs.},
 @article{gower1969,
   title={Minimum spanning trees and single linkage cluster analysis},
   author={Gower, John C and Ross, Gavin JS},
-  journal={Applied statistics},
+  journal={Applied Statistics},
   pages={54--64},
   year={1969},
   Doi                      = {10.2307/2346439},
@@ -8237,7 +8237,7 @@ Only done with MP!},
   }
 
 @Manual{RStudio, 
-  title = {RStudio: Integrated Development Environment for R}, 
+  title = {RStudio: Integrated Development Environment for {R}}, 
   author = {{RStudio Team}}, organization = {RStudio, Inc.}, 
   address = {Boston, MA}, 
   year = {2015}, 
@@ -8498,7 +8498,7 @@ Only done with MP!},
 }
 
   @Manual{shiny,
-    title = {shiny: Web Application Framework for R},
+    title = {shiny: Web Application Framework for {R}},
     author = {Winston Chang and Joe Cheng and JJ Allaire and Yihui Xie and Jonathan McPherson},
     year = {2018},
     note = {R package version 1.1.0},
@@ -10428,7 +10428,7 @@ doi = {10.1093/bioinformatics/btx155}
 @article{ohara1997,
   title={Population thinking and tree thinking in systematics},
   author={O'Hara, Robert J},
-  journal={Zoologica scripta},
+  journal={Zoologica Scripta},
   volume={26},
   number={4},
   pages={323--329},
