+++ bibliography.bib
@@ -1839,7 +1839,7 @@ Quite nice!},
 @Book{Darwin1859,
   Title                    = {On the Origin of Species by Means of Natural Selection, or the Preservation of Favoured Races in the Struggle for Life},
   Author                   = {Darwin, Charles},
-  Publisher                = {Murray},
+  Publisher                = {Murray, London},
   Year                     = {1859},
 
   Owner                    = {hoehna},
@@ -3290,7 +3290,7 @@ Uses also the diagnosis of the golden runs.},
 }
  @Book{ggplot2,
     author = {Hadley Wickham},
-    title = {ggplot2: Elegant Graphics for Data Analysis},
+    title = {{g}gplot2: Elegant Graphics for Data Analysis},
     publisher = {Springer-Verlag New York},
     year = {2016},
     isbn = {978-3-319-24277-4},
@@ -3298,7 +3298,7 @@ Uses also the diagnosis of the golden runs.},
   }
   
   @Article{ggtree,
-    title = {ggtree: an R package for visualization and annotation of phylogenetic trees with their covariates and other associated data.},
+    title = {{g}gtree: an {R} package for visualization and annotation of phylogenetic trees with their covariates and other associated data.},
     author = {Guangchuang Yu and David Smith and Huachen Zhu and Yi Guan and Tommy Tsan-Yuk Lam},
     year = {2017},
     journal = {Methods in Ecology and Evolution},
