### 2019-01-17 #35 @wrightaprilm
All right, got 'em. I also fixed a couple title case issues - R as a solo letter in an article title needs to be {R} to escape auto-format.
