### 2019-01-17 #35 @arfon
@labarba @wrightaprilm - the margin bug should also now be fixed because of the changes in https://github.com/openjournals/whedon/pull/39.
