### 2019-01-17 #35 @labarba
I just tried to compile locally with `whedon download` then `whedon compile`, and I get a similar error message. Strange that `@whedon generate pdf` on here works, but `whedon compile` fails. Aren't they supposed to be doing the same thing, @arfon?
