### 2019-01-17 #35 @labarba
My error message:

```
bundler: failed to load command: whedon (/Users/labarba/.rbenv/versions/2.5.1/lib/ruby/gems/2.5.0/bin/whedon)
NoMethodError: undefined method `has_field?' for #<BibTeX::String an = "The American Naturalist">
Did you mean?  has_type?
  /Users/labarba/GitHub/whedon/lib/whedon/bibtex.rb:44:in `make_citation'
  /Users/labarba/GitHub/whedon/lib/whedon/bibtex.rb:31:in `block in generate_citations'
  /Users/labarba/.rbenv/versions/2.5.1/lib/ruby/gems/2.5.0/gems/bibtex-ruby-4.4.7/lib/bibtex/bibliography.rb:150:in `each'
  /Users/labarba/.rbenv/versions/2.5.1/lib/ruby/gems/2.5.0/gems/bibtex-ruby-4.4.7/lib/bibtex/bibliography.rb:150:in `each'
  /Users/labarba/GitHub/whedon/lib/whedon/bibtex.rb:29:in `generate_citations'
  /Users/labarba/GitHub/whedon/lib/whedon/processor.rb:214:in `generate_crossref'
  /Users/labarba/GitHub/whedon/lib/whedon/processor.rb:91:in `compile'
  /Users/labarba/GitHub/whedon/bin/whedon:76:in `compile'
  /Users/labarba/.rbenv/versions/2.5.1/lib/ruby/gems/2.5.0/gems/thor-0.20.0/lib/thor/command.rb:27:in `run'
  /Users/labarba/.rbenv/versions/2.5.1/lib/ruby/gems/2.5.0/gems/thor-0.20.0/lib/thor/invocation.rb:126:in `invoke_command'
  /Users/labarba/.rbenv/versions/2.5.1/lib/ruby/gems/2.5.0/gems/thor-0.20.0/lib/thor.rb:387:in `dispatch'
  /Users/labarba/.rbenv/versions/2.5.1/lib/ruby/gems/2.5.0/gems/thor-0.20.0/lib/thor/base.rb:466:in `start'
  /Users/labarba/GitHub/whedon/bin/whedon:113:in `<top (required)>'
  /Users/labarba/.rbenv/versions/2.5.1/lib/ruby/gems/2.5.0/bin/whedon:23:in `load'
  /Users/labarba/.rbenv/versions/2.5.1/lib/ruby/gems/2.5.0/bin/whedon:23:in `<top (required)>'
electra:whedon labarba$
```
