### 2019-01-17 #35 @labarba
@wrightaprilm — I see that you are using a long file `bibliography.bib` as the source of citations. Would it be too annoying to copy and past just the BibTeX entries you are citing in the paper into a `paper.bib` and use that instead?
