+++ paper.bib
@@ -1,0 +1,295 @@
+@Article{abella12,
+  Title                    = {\textit{{K}retzoiarctos} gen. nov., the Oldest Member of the Giant Panda Clade},
+  Author                   = {Juan Abella and David M. Alba and Josep M. Robles and Alberto Valenciano and Cheyenn Rotgers and Ra\"ul Carmona and Plinio Montoya and Jorge Morales},
+  Journal                  = {PLoS One},
+  Year                     = {2012},
+  Pages                    = {e48985},
+  Volume                   = {17}
+}
+
+@Article{abella2011,
+  Title                    = {Una nueva especie de \textit{{A}griarctos} ({A}iluropodinae, {U}rsidae, {C}arnivora) en la localidad de {N}ombrevilla 2 ({Z}aragoza, {E}spa{\~n}a)},
+  Author                   = {Abella, J and Montoya, P and Morales, J},
+  Journal                  = {Estudios Geol{\'o}gicos},
+  Year                     = {2011},
+  Number                   = {2},
+  Pages                    = {187--191},
+  Volume                   = {67}
+}
+
+@article {Baum979,
+	author = {Baum, David A. and Smith, Stacey DeWitt and Donovan, Samuel S. S.},
+	title = {The Tree-Thinking Challenge},
+	volume = {310},
+	number = {5750},
+	pages = {979--980},
+	year = {2005},
+	doi = {10.1126/science.1117727},
+	publisher = {American Association for the Advancement of Science},
+	issn = {0036-8075},
+	URL = {http://science.sciencemag.org/content/310/5750/979},
+	eprint = {http://science.sciencemag.org/content/310/5750/979.full.pdf},
+	journal = {Science}
+}
+  @Manual{shiny,
+    title = {shiny: Web Application Framework for {R}},
+    author = {Winston Chang and Joe Cheng and JJ Allaire and Yihui Xie and Jonathan McPherson},
+    year = {2018},
+    note = {R package version 1.1.0},
+    url = {https://CRAN.R-project.org/package=shiny},
+  }
+
+@Book{Darwin1859,
+  Title                    = {On the Origin of Species by Means of Natural Selection, or the Preservation of Favoured Races in the Struggle for Life},
+  Author                   = {Darwin, Charles},
+  Publisher                = {Murray, London},
+  Year                     = {1859},
+
+  Owner                    = {hoehna},
+  Timestamp                = {2013.10.12}
+}
+
+@article{FarrisKluge,
+author = {Farris, James S. and Kluge, Arnold G. and Eckardt, Michael J.},
+title = {A Numerical Approach to Phylogenetic Systematics},
+journal = {Systematic Biology},
+volume = {19},
+number = {2},
+pages = {172-189},
+year = {1970},
+doi = {10.2307/2412452},
+URL = {http://dx.doi.org/10.2307/2412452},
+eprint = {/oup/backfile/content_public/journal/sysbio/19/2/10.2307/2412452/2/19-2-172.pdf}
+}
+
+
+@Article{Felsenstein1978,
+  Title                    = {The Number of Evolutionary Trees},
+  Author                   = {Felsenstein, J.},
+  Journal                  = {Systematic Zoology},
+  Year                     = {1978},
+  Number                   = {1},
+  DOI                      = {https://doi.org/10.2307/2412810},
+  Pages                    = {27--33},
+  Volume                   = {27},
+
[…]
