+++ paper.bib
@@ -208,7 +208,7 @@ url={https://doi.org/10.1007/s12052-010-0254-9}
   publisher={Wiley Online Library}
 }
 
-@Manual{RCoreTeam2013,
+@Manual{R,
   Title                    = {R: A Language and Environment for Statistical Computing},
 
   Address                  = {Vienna, Austria},
