### 2019-01-17 #35 @wrightaprilm
OK, had to fix one citation to R core, but looks fine. Fingers crossed!
