+++ paper.bib
@@ -124,7 +124,7 @@ eprint = {/oup/backfile/content_public/journal/sysbio/19/2/10.2307/2412452/2/19-
 @Article{Huelsenbeck2001a,
   Title                    = {{MRBAYES: Bayesian inference of phylogenetic trees}},
   Author                   = {Huelsenbeck, J.P. and Ronquist, F.},
-  Journal                  = BI,
+  Journal                  = {Bioinformatics},
   Year                     = {2001},
   Number                   = {8},
   Pages                    = {754--755},
