### 2019-01-17 #35 @labarba
Nope ... some DOI links are broken. See screenshot.
