### 2019-01-17 8530407
remove https from DOI
