+++ paper.bib
@@ -58,8 +58,6 @@ number = {2},
 pages = {172-189},
 year = {1970},
 doi = {10.2307/2412452},
-URL = {http://dx.doi.org/10.2307/2412452},
-eprint = {/oup/backfile/content_public/journal/sysbio/19/2/10.2307/2412452/2/19-2-172.pdf}
 }
 
 
@@ -69,7 +67,7 @@ eprint = {/oup/backfile/content_public/journal/sysbio/19/2/10.2307/2412452/2/19-
   Journal                  = {Systematic Zoology},
   Year                     = {1978},
   Number                   = {1},
-  DOI                      = {https://doi.org/10.2307/2412810},
+  DOI                      = {10.2307/2412810},
   Pages                    = {27--33},
   Volume                   = {27},
 
@@ -91,7 +89,7 @@ eprint = {/oup/backfile/content_public/journal/sysbio/19/2/10.2307/2412452/2/19-
   Number                   = {3},
   Pages                    = {240--249},
   Volume                   = {22},
-  DOI                      = {https://doi.org/10.1093/sysbio/22.3.240},
+  DOI                      = {10.1093/sysbio/22.3.240},
   Owner                    = {shhn001},
   Publisher                = {JSTOR},
   Timestamp                = {2008.04.02}
@@ -104,7 +102,7 @@ eprint = {/oup/backfile/content_public/journal/sysbio/19/2/10.2307/2412452/2/19-
   volume={25},
   number={4},
   pages={351--360},
-  doi = {https://doi.org/10.1007/BF02603120},
+  doi = {10.1007/BF02603120},
   year={1987},
   publisher={Springer}
 }
@@ -129,7 +127,7 @@ eprint = {/oup/backfile/content_public/journal/sysbio/19/2/10.2307/2412452/2/19-
   Number                   = {8},
   Pages                    = {754--755},
   Volume                   = {17},
-  DOI                      = {https://doi.org/10.1093/bioinformatics/17.8.754},
+  DOI                      = {10.1093/bioinformatics/17.8.754},
   Owner                    = {shhn001},
   Publisher                = {Oxford Univ Press},
   Timestamp                = {2008.04.02}
@@ -178,7 +176,7 @@ title = {College Students' Misconceptions About Evolutionary Trees},
 journal = {The American Biology Teacher},
 volume = {69},
 year = {2007},
-doi = {https://doi.org/10.1662/0002-7685(2007)69[71:csmaet]2.0.co;2}
+doi = {10.1662/0002-7685(2007)69[71:csmaet]2.0.co;2}
 }
 
 @Article{Meisel2010,
@@ -193,7 +191,7 @@ number={4},
 pages={621--628},
 issn={1936-6434},
 doi={10.1007/s12052-010-0254-9},
-url={https://doi.org/10.1007/s12052-010-0254-9}
+url={10.1007/s12052-010-0254-9}
 }
 
 @article{ohara1997,
@@ -204,7 +202,7 @@ url={https://doi.org/10.1007/s12052-010-0254-9}
   number={4},
   pages={323--329},
   year={1997},
-  doi={https://doi.org/10.1111/j.1463-6409.1997.tb00422.x},
+  doi={10.1111/j.1463-6409.1997.tb00422.x},
   publisher={Wiley Online Library}
 }
 
@@ -237,7 +235,7 @@ url={https://doi.org/10.1007/s12052-010-0254-9}
   number={10},
   pages={1069--1089},
   year={1998},
-  DOI={https://doi.org/10.1002/(SICI)1098-2736(199812)35:10<1069::AID-TEA2>3.0.CO;2-A},
+  DOI={10.1002/(SICI)1098-2736(199812)35:10<1069::AID-TEA2>3.0.CO;2-A},
   publisher={Wiley Online Library}
 }
 
