### 2019-01-17 #35 @arfon
I think bibtex expects `doi` fields in `paper.bib` to be of the form `10.1662/0002-7685(2007)69[71:csmaet]2.0.co;2` not `https://doi.org/10.1662/0002-7685(2007)69[71:csmaet]2.0.co;2`. Entries like this one need updating:
https://github.com/wrightaprilm/treesiftr/blob/master/paper.bib#L181
