### 2019-01-17 #35 @wrightaprilm
Right, just went ahead and stripped off the `https' wrappers. Will re-render.
