### 2019-01-17 #35 @labarba
Checking all the links and ... http://ggplot2.org shows "not found"?
