+++ paper.bib
@@ -266,7 +266,7 @@ doi = {10.1093/bioinformatics/btx155}
     publisher = {Springer-Verlag New York},
     year = {2016},
     isbn = {978-3-319-24277-4},
-    url = {http://ggplot2.org},
+    url = {https://ggplot2.tidyverse.org},
   }
   
   @Article{ggtree,
