### 2019-01-17 #35 @wrightaprilm
I changed the link in my .bib. That's from the maintainer's own bibtex entry on their site. I should probably make a pull request and let them know its broken. Good catch!
