### 2019-01-17 #35 @labarba
Congratuations, @wrightaprilm, your paper is published! 🎉
(you made us work hard for this one, phew!)

Big thank you to your handling editor, @juanklopper
and your reviewers, @ethanwhite and @rachelss
🙏
